\name{HPOSim-internal}
\title{Internal HPOSim objects}
\alias{SaveGenesToPhenotype}
\alias{calculateICS}
\alias{getAncestors}
\alias{getChildren}
\alias{getGeneDefaultBackground}
\alias{getDiseaseDefaultBackground}
\alias{getGIC}
\alias{getGSim}
\alias{getDIC}
\alias{getDSim}
\alias{getOffsprings}
\alias{getParents}
\alias{getTerms}
\alias{initialize}
\alias{normalize.kernel}
\alias{hpo2gene}
\alias{gene2hpo}
\alias{hpo2disease}
\alias{disease2hpo}
\alias{termIC}
\alias{DiseasetermIC}
\description{Internal HPOSim objects.}
\details{These are not to be called by the user.}
\keyword{internal}
