\name{HPOSim-internal}
\title{Internal HPOSim objects}
\alias{calculateGeneTermIC}
\alias{getAncestors}
\alias{getChildren}
\alias{getGeneDefaultBackground}
\alias{getDiseaseDefaultBackground}
\alias{getGIC}
\alias{getOffsprings}
\alias{getParents}
\alias{getTerms}
\alias{getLCA}
\alias{dp}
\alias{fmax}
\alias{calculateDiseaseTermIC}
\alias{getTermOffspringDiseases}
\alias{RemoveTermsWithoutIC}
\alias{.initialize}
\alias{.initial}
\alias{normalize.kernel}
\alias{hpo2gene}
\alias{gene2hpo}
\alias{hpo2disease}
\alias{disease2hpo}
\alias{termIC}
\alias{DiseasetermIC}
\description{Internal HPOSim objects.}
\details{These are not to be called by the user.}
\keyword{internal}
