\name{HPOSim-package}
\alias{HPOSim-package}
\alias{HPOSim}
\docType{package}
\title{ HPOSim package}
\description{
  This package implements multiple similarity measures and enrichment analysis functions for HPO terms ,gene products and diseases.
}
\details{
  \tabular{ll}{
    Package: \tab HPOSim\cr
    Type: \tab Package\cr
    Version: \tab 1.0\cr
    Date: \tab 2013-03-25\cr
    License: \tab GPL-2\cr
  }
}
\author{
  Yue Deng, Gang Wang, Xiaocheng Huang and Tingting Ma
  Maintainer: Yue Deng <anfdeng@163.com>
}
\keyword{ package }
